抜けていた例外処理の追加。 正規表現を用いたコードの判別の追加。 合計金額をIntではなくLongで扱うようにして10桁越えた際のエラーを避けるようにした。

+++ arraytypeAggregateSales/allData.java
@@ -2,9 +2,9 @@ package arraytypeAggregateSales;
 
 public class allData {
 	String name;
-	int sales;
+	long sales;
 
-	protected allData(String name , int sales){
+	protected allData(String name , long sales){
 		this.name = name;
 		this.sales = sales;
 	}
